+++ include/cmd_array.h
@@ -12,6 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "parse_line.h"
+#include <string.h>
+#include <glob.h>
+#include "exec_cmd.h"
 
-int main(int argc, char **argv);
+enum {
+	MAX_COMMANDS_PER_LINE = 32
+};
+
+struct cmd_array {
+	struct command *commands[MAX_COMMANDS_PER_LINE];
+	int n_cmd;
+	int status;
+};
+
+struct cmd_array *new_commands();
+
+int add_command(struct command *command_to_add, struct cmd_array *commands);
+
+int do_not_wait_commands(struct cmd_array *cmd_array);
+
+void free_cmd_array(struct cmd_array *cmd_array);
